--- a/spec/features/layout/footer_spec.rb
+++ b/spec/features/layout/footer_spec.rb
@@ -5,35 +5,28 @@ require "rails_helper"
 RSpec.describe "Footer displays" do
   before { visit login_path }
 
-  it "contact" do
-    within("footer") { expect(page).to have_content("Contact") }
-  end
+  it "contact details" do
+    within("footer") do
+      expect(page).to have_content("Contact")
 
-  describe "contact of" do
-    before { click_on "Contact" }
-
-    it "email" do
-      within("footer") { expect(page).to have_link(nil, href: "mailto:juzershakir.webdev@gmail.com") }
-    end
-
-    it "whatsapp" do
-      within("footer") { expect(page).to have_link(nil, href: "https://wa.me/919819393148") }
-    end
-
-    it "telegram" do
-      within("footer") { expect(page).to have_link(nil, href: "https://t.me/juzershakir") }
+      click_on "Contact"
+      # Email
+      expect(page).to have_link(nil, href: "mailto:juzershakir.webdev@gmail.com")
+      # WhatsApp
+      expect(page).to have_link(nil, href: "https://wa.me/919819393148")
+      # Telegram
+      expect(page).to have_link(nil, href: "https://t.me/juzershakir")
     end
   end
 
-  it "about" do
-    within("footer") { expect(page).to have_link("About", href: "https://www.thedawoodibohras.com/about-the-bohras/faiz-al-mawaid-al-burhaniyah/") }
-  end
-
-  it "dynamic year" do
-    within("footer") { expect(page).to have_content(Date.current.year) }
-  end
-
-  it "creator name linking to its portfolio site" do
-    within("footer") { expect(page).to have_link("Juzer Shakir", href: "https://juzershakir.github.io/") }
+  it "Other Citations" do
+    within "#footer" do
+      # About
+      expect(page).to have_link("About", href: "https://www.thedawoodibohras.com/about-the-bohras/faiz-al-mawaid-al-burhaniyah/")
+      # Year
+      expect(page).to have_content(Date.current.year)
+      # creator name linking to its portfolio site
+      expect(page).to have_link("Juzer Shakir", href: "https://juzershakir.github.io/")
+    end
   end
 end