--- a/spec/features/layout/navbar_spec.rb
+++ b/spec/features/layout/navbar_spec.rb
@@ -15,63 +15,43 @@ RSpec.describe "Navbar" do
 
       # * Statistics
       describe "Statistics dropdown menu" do
-        it do
-          within("#statistics") { expect(page).to have_link("Sabeels", href: statistics_sabeels_path) }
-        end
-
-        it do
-          within("#statistics") { expect(page).to have_link("Thaalis", href: statistics_thaalis_path) }
+        it "has links to Sabeels & Thaalis" do
+          within("#statistics") do
+            expect(page).to have_link("Sabeels", href: statistics_sabeels_path)
+            expect(page).to have_link("Thaalis", href: statistics_thaalis_path)
+          end
         end
       end
 
       # * Resources
       describe "Resources dropdown menu" do
-        it do
-          within("#resources") { expect(page).to have_link("Sabeels", href: sabeels_path) }
-        end
-
-        it do
-          within("#resources") { expect(page).to have_link("Thaalis", href: thaalis_all_path(CURR_YR)) }
-        end
-
-        it do
-          within("#resources") { expect(page).to have_link("Transactions", href: transactions_all_path) }
+        it "has links to Sabeels, Thaalis & Transactions" do
+          within("#resources") do
+            expect(page).to have_link("Sabeels", href: sabeels_path)
+            expect(page).to have_link("Thaalis", href: thaalis_all_path(CURR_YR))
+            expect(page).to have_link("Transactions", href: transactions_all_path)
+          end
         end
       end
     end
 
     # * Admin
     context "when admin" do
-      describe "has logged in, it displays" do
-        let(:user) { create(:admin_user) }
+      let(:user) { create(:admin_user) }
 
-        it do
-          within(".navbar-nav") { expect(page).to have_css("#admin", text: "Admin") }
+      it "Show 'Admin' dropdown & Create Sabeel button" do
+        within(".navbar-nav") do
+          expect(page).to have_css("#admin", text: "Admin")
+          expect(page).to have_link("Create Sabeel", href: new_sabeel_path)
         end
+      end
 
-        # * New Sabeel
-        it do
-          within(".navbar-nav") { expect(page).to have_link("Create Sabeel", href: new_sabeel_path) }
-        end
-
-        # * My Profile
-        it do
-          within("#admin") { expect(page).to have_link("My Profile", href: user_path(user)) }
-        end
-
-        # * New User
-        it do
-          within("#admin") { expect(page).to have_link("New User", href: new_user_path) }
-        end
-
-        # * Home
-        it do
-          within("#admin") { expect(page).to have_link("All Users", href: users_path) }
-        end
-
-        # * Log Out
-        it do
-          within("#admin") { expect(page).to have_link("Log out", href: destroy_path) }
+      it "'Admin' dropdown" do
+        within("#admin") do
+          expect(page).to have_link("My Profile", href: user_path(user))
+          expect(page).to have_link("New User", href: new_user_path)
+          expect(page).to have_link("All Users", href: users_path)
+          expect(page).to have_link("Log out", href: destroy_path)
         end
       end
     end
@@ -80,71 +60,42 @@ RSpec.describe "Navbar" do
     context "when Member" do
       let(:user) { create(:member_user) }
 
-      describe "it displays" do
-        it do
-          within(".navbar-nav") { expect(page).to have_css("#member", text: "Member") }
-        end
-
-        # * My Profile
-        it do
-          within("#member") { expect(page).to have_link("My Profile", href: user_path(user)) }
-        end
-
-        # * Log Out
-        it do
-          within("#member") { expect(page).to have_link("Log out", href: destroy_path) }
-        end
+      it "Show 'Member' dropdown" do
+        within(".navbar-nav") { expect(page).to have_css("#member", text: "Member") }
       end
 
-      describe "doesn't display" do
-        # * Home
-        it do
-          within("#member") { expect(page).to have_no_content("Home") }
+      it "'Member' dropdown" do
+        within("#member") do
+          expect(page).to have_link("My Profile", href: user_path(user))
+          expect(page).to have_link("Log out", href: destroy_path)
         end
-      end
-
-      # * New User
-      it do
-        within("#member") { expect(page).to have_no_content("New User") }
       end
     end
 
     # * Viewer
-    context "when Viewer, it displays" do
+    context "when Viewer" do
       let(:user) { create(:viewer_user) }
 
-      it "have 'viewer' id" do
+      it "has 'viewer' id" do
         within(".navbar-nav") { expect(page).to have_css("#viewer") }
       end
 
-      # * Log Out
-      it do
-        within("#viewer") { expect(page).to have_link("Log out", href: destroy_path) }
-      end
-
-      # * My Profile
-      it do
-        within("#viewer") { expect(page).to have_no_content("My Profile") }
-      end
-
-      # * Home
-      it do
-        within("#viewer") { expect(page).to have_no_content("Home") }
-      end
-
-      # * New User
-      it do
-        within("#viewer") { expect(page).to have_no_content("New User") }
+      it "within 'viewer' id" do
+        expect(page).to have_link("Log out", href: destroy_path)
+        expect(page).to have_no_content("My Profile")
       end
     end
 
     # * Member or Viewer
-    describe "Member or Viewer, do NOT display" do
+    context "when Member or Viewer" do
       let(:user) { create(:user_member_or_viewer) }
 
-      # * New Sabeel
-      it do
-        within(".navbar-nav") { expect(page).to have_no_content("Create Sabeel") }
+      it "within the navbar" do
+        within(".navbar-nav") do
+          expect(page).to have_no_content("Create Sabeel")
+          expect(page).to have_no_content("Home")
+          expect(page).to have_no_content("New User")
+        end
       end
     end
   end