--- a/spec/features/pages/home_spec.rb
+++ b/spec/features/pages/home_spec.rb
@@ -5,20 +5,20 @@ require "rails_helper"
 RSpec.describe "Page Home template, displays" do
   before { visit root_path }
 
-  it "logo" do
-    within("#home__header") { expect(page).to have_css("img[src*='logos/fmb']") }
+  describe "header" do
+    it "has a logo" do
+      within("#home__header") { expect(page).to have_css("img[src*='logos/fmb']") }
+    end
   end
 
-  it "secondary header" do
-    within("#home__body") { expect(page).to have_css("h5") }
-  end
-
-  it "primary header" do
-    within("#home__body") { expect(page).to have_css("h2") }
-  end
-
-  it "CTA button" do
-    within("#home__body") { expect(page).to have_link("Free demo", href: login_path) }
+  describe "body" do
+    it "has header tags & CTA button" do
+      within("#home__body") do
+        expect(page).to have_css("h5")
+        expect(page).to have_css("h2")
+        expect(page).to have_link("Free demo", href: login_path)
+      end
+    end
   end
 
   it { expect(page).to have_css("#footer") }