--- a/spec/features/sabeels/active_spec.rb
+++ b/spec/features/sabeels/active_spec.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require "rails_helper"
-require_relative "sabeel_helpers"
 
 RSpec.describe "Sabeel Active template" do
   let(:apt) { "Burhani" }
@@ -14,27 +13,27 @@ RSpec.describe "Sabeel Active template" do
   end
 
   # * ALL user types
-  describe "visited by any user type", :js do
-    it { expect(page).to have_title "Active Sabeels - #{apt}" }
+  describe "visited by any user type" do
+    it "has a title and Generate PDF button" do
+      expect(page).to have_title "Active Sabeels - #{apt}"
+      expect(page).to have_link("Generate PDF")
+      expect(page).to have_css(".fa-file-pdf")
+    end
 
-    describe "Generate PDF button" do
+    describe "Generates PDF of an apartment", :js do # rubocop:disable RSpec/MultipleMemoizedHelpers
       let(:sabeel) { sabeels.first }
       let(:thaali) { sabeel.thaalis.first }
+      let(:pdf) { switch_to_window(windows.last) }
 
-      it { expect(page).to have_link("Generate PDF") }
-      it { expect(page).to have_css(".fa-file-pdf") }
+      before { click_on("Generate PDF") }
 
-      describe "Generates PDF of an apartment" do # rubocop:disable RSpec/MultipleMemoizedHelpers
-        let(:pdf_window) { switch_to_window(windows.last) }
-
-        before { click_on("Generate PDF") }
-
-        it { within_window pdf_window { expect(page).to have_content("#{apt} - #{CURR_YR}") } }
-        it { within_window pdf_window { expect(page).to have_content(sabeel.flat_no) } }
-        it { within_window pdf_window { expect(page).to have_content(sabeel.name) } }
-        it { within_window pdf_window { expect(page).to have_content(sabeel.mobile) } }
-        it { within_window pdf_window { expect(page).to have_content(thaali.number) } }
-        it { within_window pdf_window { expect(page).to have_content(thaali.size.humanize.chr) } }
+      it do
+        within_window pdf { expect(page).to have_content("#{apt} - #{CURR_YR}") }
+        within_window pdf { expect(page).to have_content(sabeel.flat_no) }
+        within_window pdf { expect(page).to have_content(sabeel.name) }
+        within_window pdf { expect(page).to have_content(sabeel.mobile) }
+        within_window pdf { expect(page).to have_content(thaali.number) }
+        within_window pdf { expect(page).to have_content(thaali.size.humanize.chr) }
       end
     end