--- a/spec/features/sabeels/sabeel_helpers.rb
+++ b/spec/features/sabeels/sabeel_helpers.rb
@@ -1,9 +1,0 @@
-shared_examples "view sabeel records" do
-  it "name" do
-    sabeels.each { expect(page).to have_content(_1.name) }
-  end
-
-  it "apartment" do
-    sabeels.each { expect(page).to have_content(_1.apartment.titleize) }
-  end
-end