--- a/spec/features/thaalis/complete_spec.rb
+++ b/spec/features/thaalis/complete_spec.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require "rails_helper"
-require_relative "thaali_helpers"
 
 RSpec.describe "Thaali complete template" do
   let(:user) { create(:user) }
@@ -10,12 +9,11 @@ RSpec.describe "Thaali complete template" do
   before do
     sign_in(user)
     create_list(:taking_thaali_dues_cleared, 2)
-
     visit thaalis_complete_path(CURR_YR)
   end
 
   # * ALL user types
-  describe "visited by any user type can", :js do
+  describe "visited by any user type can" do
     it { expect(page).to have_title "Completed Takhmeens in #{CURR_YR}" }
 
     it_behaves_like "view thaali records"