… .rubocop.yml
… spec/features/shared_helpers.rb
… spec/features/thaalis/destroy_spec.rb
… spec/features/thaalis/new_spec.rb
… spec/features/thaalis/pending_spec.rb
… spec/features/thaalis/show_spec.rb
… spec/features/thaalis/thaali_helpers.rb
… spec/features/transactions/all_spec.rb
… spec/features/transactions/destroy_spec.rb
… spec/features/transactions/edit_spec.rb
… spec/features/transactions/new_spec.rb
… spec/features/transactions/show_spec.rb
… spec/features/transactions/transaction_helpers.rb
… spec/features/users/destroy_spec.rb
… spec/features/users/edit_spec.rb
… spec/features/users/index_spec.rb
… spec/features/users/new_spec.rb
… spec/features/users/show_spec.rb
… spec/support/matchers/have_humanized_number_matcher.rb
… spec/support/shared_examples.rb