This PR has significantly improved the RSpec Feature Tests by merging similar test suites into one, improving readability & performance.

| Metric | Before PR (secs) | After PR (secs) |
|--------|--------|--------|
| Execution Time (seconds) | 55-60 | **20-25**  |
| Number of Examples | 521 | **414** |
| Improvement in Execution Time |  | **61%** |
| Improvement in Number of Examples |  | **20.5%** |

- Create a custom RSpec Matcher `have_humanized_number` to test against the `number_to_human` function. This refactored repeated logic resulting in better readability, lower probability of errors, and code elegance.